requirements.txt
pandas
numpy
scipy
statsmodels
plotly
seaborn
streamlit
fpdf2
matplotlib